requirements.txt
transformers==4.39.3
torch==2.0.0
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
Flask==3.0.3
pandas
scikit-learn